# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The backend of the fibonacci spiral matrix implementation",
        "title": "Spiral Matrix API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Verifies users credentials and generate a JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login a user",
                "parameters": [
                    {
                        "description": "User Credentials",
                        "name": "models.AuthUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_models.AuthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_models.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            }
        },
        "/public/user": {
            "post": {
                "description": "Store a new User in the database, this endpoint doesn't need authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a User",
                "parameters": [
                    {
                        "description": "Add User",
                        "name": "models.CreateUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_models.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            }
        },
        "/spiral": {
            "get": {
                "description": "Generates a Matrix of size cols x rows filled with fibonacci numbers and ordered like a spiral",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spiral"
                ],
                "summary": "Generate Fibonacci Spiral",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "columns",
                        "name": "cols",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "rows",
                        "name": "rows",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "List all non deleted users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/spiralmatrix_internal_models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft deletes a User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Deletes a User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Get user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "description": "Retrieves a non deleted user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieves a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Get user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            },
            "patch": {
                "description": "Verifies the current password and then changes the password of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Change Password of user",
                "parameters": [
                    {
                        "description": "Change Password",
                        "name": "models.ChangePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_models.ChangePassword"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Get user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/spiralmatrix_internal_utils.ErrorWrapper"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "spiralmatrix_internal_models.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/spiralmatrix_internal_models.User"
                }
            }
        },
        "spiralmatrix_internal_models.AuthUser": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "spiralmatrix_internal_models.ChangePassword": {
            "type": "object",
            "required": [
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        },
        "spiralmatrix_internal_models.CreateUser": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "spiralmatrix_internal_models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "spiralmatrix_internal_utils.ErrorWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}